requirements.txt
python-dotenv>=1.0,<2
pydantic>=2.0,<3
openai>=2.0,<3
modal>=1.0,<2
fastapi[standard]>=0.100,<1
uvicorn>=0.30,<1
pytest>=8.0,<9
httpx>=0.27,<1
